+++ Model/DTO/PaginatedGenericDto.cs
@@ -6,6 +6,6 @@
         public int PageSize { get; set; }
         public int TotalPages { get; set; }
         public T Result { get; set; }
-        public long TotalMinuteUsed { get; set; }
+        public decimal TotalMinuteUsed { get; set; }
     }
 }
