+++ Model/DTO/UserTimeUsedDto.cs
@@ -12,6 +12,6 @@ namespace Model.DTO
         public string Email { get; set; }
         public string PhoneNumber { get; set; }
         public string UserName { get; set; }
-        public int TotalTimeUsed { get; set; }
+        public decimal TotalTimeUsed { get; set; }
     }
 }
